+++ make_package.R
@@ -11,9 +11,9 @@ library(devtools)
 
 setwd('~/github/nimble/nimbleHMC')
 
-document('nimbleHMC')
+devtools::document('nimbleHMC')
 
-system('R CMD BUILD nimbleHMC')
+devtools::build('nimbleHMC')
 
 check('nimbleHMC')
 
@@ -22,6 +22,8 @@ suppressMessages(try(remove.packages('nimbleHMC'), silent = TRUE))
 (lastTarFile <- tarFiles[length(tarFiles)])
 message('installing package version ', gsub('\\.tar\\.gz$', '', lastTarFile))
 system(paste0('R CMD install ', lastTarFile))
+
+##devtools::install('nimbleHMC')
 
 q('no')