+++ nimbleHMC/inst/NEWS.md
@@ -1,3 +1,11 @@
+#                            CHANGES IN VERSION 0.2.1 (February 2024)
+
+- Added checks within `NUTS` and `NUTS_classic` samplers to check validity of target nodes.
+
+- Testing of correct results of HMC samplers on CAR model structures.
+
+- `NUTS` and `NUTS_classic` samplers updated consistent with version 1.1.0 of `nimble`.
+
 #                            CHANGES IN VERSION 0.2.0 (September 2023)
 
 - Renamed `HMC` sampler to `NUTS_classic`, and added `NUTS` HMC sampler which follows the implementation of version 2.32.2 of Stan.